requirements.txt
Mezzanine==3.1.10
django-tastypie
djangorestframework